Bug 481796 - IndexBinaryType bugfixes and refactoring
Fixed a number of return types which should have returned null but were
returning empty arrays in error situations.

Wrote a new TypeRef object to use instead of DatabaseRef<NdType>, and
adopted it. The extra generality of the generic version was making it
harder and more verbose to use directly.

Change-Id: I8a2976d90232bf33bf4aaa887facb62a6c5b3188
Signed-off-by: Stefan Xenos <sxenos@gmail.com>