Bug 526591 - fix "using .equals to compare two char[]'s"
Calling equals() on an array is the same as comparing their addresses,
but it does not compare the array contents and could lead to NPE's if
array reference was null. Sometimes the intent was really to compare the
contents, so the code never worked in the expected way.

Avoid confusing or wrong code and potential NPE's by explicitly check
pointer equality using == or by using Arrays.equals() to compare the
array contents.

Change-Id: I08f774657ac1be54ad4895764d4f7168b440b630
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>