Bug 526010 - Improved traversal of empty parents on package delete
When removing lots of packages, for each package JavaModelOperation will
go up the parent structure and try to remove empty packages. This is
very IO intensive (i.e. slow).

With this change parents are traversed only as long as they are empty.
Since deleting package/sources is done bottom-up, this spares some IO.

Change-Id: I1f96c52b8bfc8df66c4a6fc628595235cf05309c
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>