Bug 542860 - Application hangs at JavaRuntime.getDefaultVMInstall()
The fix for bug 252571 introduces a potential deadlock on an application
start-up, as soon as the application tries to access JavaCore
functionality.

In particular, an auto-build or a refresh job will hold a workspace lock
and also access JavaCore functionality. The application will then
attempt to open the external folders project due to the fix for bug
252571, as part of the JavaCore start-up. So the application (i.e. the
main thread) will require a project lock, while blocking access to
JavaCore. The auto-build or refresh job must wait for JavaCore to become
accessible, but will not release the workspace lock they hold. This
results in a deadlock.

This change prevents the deadlock from occurring, by preparing the
external folders in a separate job, at JavaCore start-up. Opening the
external folders project therefore does not occur on JavaCore start-up.

Change-Id: Idedd2157c9e3c9c4edbcca6022763c1abcc09bec
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>