Bug 545491 - revert ReferenceCollection fix
The fix caused build state to be persisted or restored in a wrong way,
so that ReferenceCollection was not properly read form a previous state,
so that types referenced by that collection were not found by Java
search anymore till next build of the project.

This change also decrements State.VERSION counter to trigger a rebuild
and a new state creation.

Tests aren't reverted, only those depending on the sorted state are
commented out.

Change-Id: I945dd6ef42eacfe34f06361b41697c7b8e7b6725
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>