Bug 548456 - Ensure core model tests fail if test case misses suite()
org.eclipse.jdt.core.tests.model.AllJavaModelTests will visit each of
the included test cases and call its suite() method over reflection to
instantiate the test case. If the suite() method is not found, an
exception stack trace is printed instead of throwing an exception. So
forgetting the suite() method in a new test case does not result in a
failure. This can easily hide the fact that the new test case is not
running at all.

This change re-throws the exception instead of only printing its stack
trace, so that missing suite() method in an included test case results
in a failure.

Change-Id: I2e8cc25cca943961caf815cd9bc04eaf781facd4
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>