Bug 564922 - Compiler fails to use Java 16 when release option is enabled
Changed "hex" release number parsing for ct.sym release directories to
"special" encoding.

Basically it *looks like* the releases starting with Java 10 encoded
with capital letters starting with "A". We assumed that this would mean
hex encoding, but at least Java 16 says that 16 should be spelled like
"G", and not like "10". Let's hope that there is some logic and 17 will
be "H" and not "$%&".

Change-Id: I5f1cccbd128898067eecf34a059674ad7aa3cea1
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>