Bug 566498 - Optimize classpath scanning to selected packages only
The existing implementation had an O(n) performance where n directly
depends on classpath size. Thus, the lookup performance increased
directly with the numbers of jars and source folders. The new
implementation narrows the scope down to package fragments first. This
is achieved by maintaining an index for package names.

This change includes a test for JavaSearchNameEnvironment which
documents some hidden implementation detail required for the index to
work properly.


Change-Id: I3bdba84f8aeffbc92b0de3ac86e6e445d906ebb2