Bug 570078 - Add new index to narrow the search index collection
The new meta index will keep track of index names and their type
reference information. So that it can be used when searching super type
references to narrow the number of indexes that been searched. More
information of this solution can be found at
https://wiki.eclipse.org/JDT_Core_Programmer_Guide/MetaIndex

This feature can be disabled by adding
-Dorg.eclipse.jdt.disableMetaIndex=false into eclipse vm arguments.

Change-Id: I6cfc76ac59506af612d5cb3c57ebafbe791b7b46
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/178816
Tested-by: JDT Bot <jdt-bot@eclipse.org>