Bug 572393 - fix Intermittent random junit gerrit failures in jdt.core
- GenericTypeTest: currentTimeMillis() is not a valid measure on a build
machine that is used for several builds in parallel. Hopefully used cpu
time is a better measure (accuracy is the OS scheduler timeslice).

- CompletionTests2: the monitor cancellation delay of one second is too
unreliable for the build machines we have today, use one minute instead
(the test itself will only wait as much as needed up to the limit, but
not necessarily a minute).


Change-Id: I45957c31957bef3aca425e21f6a44622be18aabd
Signed-off-by: Joerg Kubitz <jkubitz-eclipse@gmx.de>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/178550
Tested-by: Andrey Loskutov <loskutov@gmx.de>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>