Bug 573153: Fix whitespace issues in compiler usage texts
First and foremost, 'misc.usage.warn' contained tabs instead of spaces,
making assuming a tab width of 4 inside the properties file. But on the
console usually tabs are displayed as 8 spaces. As this is
user-configurable, though, it makes sense to expand the tabs. BTW, not
even when setting the console tab width to 4, output looks correct,
because in the file each continuation line begins with a '\' character,
making the first tab expand to position 4 in the file, but position 3 on
the output line because the backslash is not printed. This makes using
tabs unreliable at best, when trying to predict what the usage text
might look like on the console. Hence, all tabs were expanded to spaces.

In a few other places where indentation for the column layout was also
broken using normal using normal space, I made a few more cosmetic
fixes.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>
Change-Id: Icc25a1e2384c9f22a89a4480c8bf3d39d4ac81aa
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/180757
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Jay Arthanareeswaran <jarthana@in.ibm.com>