Bug 573105 - Support member completions inside lambda method invocations
When invoking completions of variable type members inside a method
invocation which is in a lambda expression with completing the whole
statement without a semicolon, it fails to resolve the binding for
completion node. This is because the outer expression which is the
invocation which contains lambda is not provided in the recovered
statements.

To fix this we provide more context as outerExpression in the
CompletionNodeDetector whenever we detect an inner expression that is a
potential poly expression.

Change-Id: If4c9ea5fc1b6c284186c2c7b1a2bfcae08cff914
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Also-by: Stephan Herrmann <stephan.herrmann@berlin.de>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/179807
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Stephan Herrmann <stephan.herrmann@berlin.de>