Bug 573313 - Fix method parameter completions at any parameter index
Before fix method parameter completions resolved to correct parameter
type only if the method parameters are filled one by one. This is
because how the expected types were picked from the method parameters.

The fix use the index of the argument which the completion happens and
use that to pick the method parameter expected type.


Change-Id: If0227457d9860ab7c456336e72546157b3d68d4b
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Also-by: Stephan Herrmann <stephan.herrmann@berlin.de>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/180143
Reviewed-by: Stephan Herrmann <stephan.herrmann@berlin.de>
Tested-by: JDT Bot <jdt-bot@eclipse.org>