Bug 571785 - Eclipse compiled code throws java.lang.ClassFormatError "Class name is empty or contains illegal character in descriptor in class file"
In this case we are creating a captureBinding with explicit binding for
the given code
public void update(final Collection<? extends byte[]> inputs) {
		inputs.forEach(this::update);
	}

According to jls-4.5.1 Collection<? extends byte[]> is valid code

Here we need to generate signature for update(capture#1-of ? extends
byte[]) for the line this::update

According the jvms-4.7.9.1 we should generate method type signature as
([B)V
Bug we are generating
(L[B;)V
Here jvm is expecting a referencetype after L but we are providing
arrayType. this is resulting in class format exception.


Change-Id: I308b7934349e4404049ab4ab4518505dc4cdfc00
Signed-off-by: Sravan Kumar Lakkimsetti <sravankumarl@in.ibm.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/180194
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Stephan Herrmann <stephan.herrmann@berlin.de>