Bug 573589 - Add support for lambda evaluation with static imports
When performing lambda expression evaluation where the source file
contains static imports, the static imports are added as normal imports
into the code snippet. This happens both for source types and binary
types with source attached. For binary types with source attached this
is fixed in the SourceMapper where the imports are captured.

Change-Id: I5ce337191d688138e32f9c43adbbe652851ec90b
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/180745
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Sarika Sinha <sarika.sinha@in.ibm.com>