Bug 573789 - Fix first method completions on a allocation expression.
When there is a complete method chaining expression which start with a
allocation expression, when invoking type completion on the first method
after the allocation expression, it fails to report the methods found on
the type of allocation expression because methods are filtered out
because when the allocation expression is a receiver it was searching
only for static methods on the type.

With the fix now allocation expressions are no longer searching only for
static methods.

Change-Id: I5addcc4202beb07694bcd1d9af643fcb815116fa
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/181155
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Stephan Herrmann <stephan.herrmann@berlin.de>