Bug 574171 - Fix deadlock when updating meta index
When meta index is tried to update from another thread through a
synchronized method on IndexManager, while the indexing thread which
holds the meta index RW lock tries to call a synchronize method on
IndexManager, they end up in a deadlock. This happens in situations
where files are requested to be removed from index.

To avoid this the fix tries to push such meta index updates which are
coming from non indexing threads into the indexing thread and run as
soon as possible to avoid ordering issues.

Change-Id: Ie7af1efd2f908f6f06c76f4b817e53247f2e95f7
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/181878
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>