Bug 573943 - [17][switch pattern][search] Search changes for JEP 406
There are 2 main changes done here. 

First is the addition of local variable declaration of the case
statement for SelectionParser in buildInitialRecoveryState.
Second is the searching for the type pattern variable of CaseStatement
in MatchLocatorParser 
Both the changes are required in selecting a pattern variable in case
statement as well as looking for the declaration of the same.

Change-Id: Ie453a6609ccf52d629dfbe034f2db79015dd1e5e
Signed-off-by: Vikas Chandra <Vikas.Chandra@in.ibm.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/182644