Bug 574532 - [dogfooding][cleanup] Use new StringBuffer -> StringBuilder cleanup for tests.compiler
To ensure that the new JDT cleanup works correctly for clients we should
run it also on our own code base. This help to ensure that the cleanup
works fine and helps us to cleanup our own code base.

From the commit message of the cleanup:

- create a cleanup to replace usage of StringBuffer with
  StringBuilder.  The two are equivalent, but StringBuffer is
  thread-safe and synchronized which makes it slower than
  StringBuilder which was introduced in Java 1.5.


To avoid API changes, we should only use the local variable change.

Change-Id: I6e4fc2164e0a5132bbb643869d6e50b89560e817
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/182648
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>