Bug 563020 - Fix chain completion with overloaded method references
When using chain completions such as stream operations with method
references of overloaded methods as parameters, earlier it failed
resolve the type bounds based on method reference because the method
declaration which contains the statements are marked as has syntax
errors.

The fix try to ignore the syntax errors at CompletionParser after
parsing block statements, if the assistNode represents the node at
cursor.


Change-Id: I39e664964cf89518563eda5eb8e7efc4b10b7ed7
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/182718
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Stephan Herrmann <stephan.herrmann@berlin.de>