Bug 574532 - [dogfooding][cleanup] Use new StringBuffer -> StringBuilder cleanup for tests.model
To ensure that the new JDT cleanup works correctly for clients we should
run it also on our own code base. This help to ensure that the cleanup
works fine and helps us to cleanup our own code base.

From the commit message of the cleanup:

- create a cleanup to replace usage of StringBuffer with
  StringBuilder.  The two are equivalent, but StringBuffer is
  thread-safe and synchronized which makes it slower than
  StringBuilder which was introduced in Java 1.5.


To avoid API changes, we should only use the local variable change.

Change-Id: Ie6052b3cbf2321602d7e7ce9b0c2dc30beb0e276
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/182668
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>