Bug 577031 - Standalone ECJ fails to build from source
ECJ now uses syntax from Java 9+ and fails to build with
source/target=1.8, this change sets instead release=11 to match the
BREE, which fixes build errors complaining about invalid Java 8 syntax.

To support records etc, ECJ requires Java 16/17 APIs, but it should
still build and run on Java 11, as indicated by the BREE. Include Java
17 API stubs in the build time classpath to fix symbol not found errors
when ECJ is building built on Java 11.

Set debug=true on the javac task so that the requested debug information
is actually generated in the bytecode. Previously the "debuglevel"
parameter was totally ignored.

Signed-off-by: Mat Booth <mat.booth@gmail.com>
Change-Id: I3669a74adc33aedfbdacb30da64fbe3baa76cdb1
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/187265
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>