Bug 577885 - Fix completions on chained method  arguments
When trying to find argument completions for chained methods which is
followed by a method with a method-ref argument, the reference
expression fails to resolve bindings due to the syntax error.

The fix extends the previous fix given to a similar problem in reference
expression resolution trying to extend support to resolve reference
expression by removing the syntax error flag from the method block.

Change-Id: I72e333fca002b130dca399bcea9a25e2abb5b5af
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/189089
Tested-by: JDT Bot <jdt-bot@eclipse.org>