Bug 578167 - fix completion for arguments inside lambda
When a method CompletionOnMessageSend is inside a lambda block, The
first expected types computations seems to empty the expected types
array, which should be in length of 1. This cause later computations to
find expected types for current argument position fails with an AIOB.
The fix guard the array copy which erroneously resize the expected types
array to zero length in first computation.


Change-Id: I34bc9a6409d8c6623d3de008764b1cd93a2520ca
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/189486
Tested-by: JDT Bot <jdt-bot@eclipse.org>