Bug 566155: implicit lambda compilation
Constructor reference was not compiled as implicit lambda when it
depends on outer instance. It does not harm code execution, but this
lambda cannot be deserialized.

When constructor is invoked without method reference, it generates an
explicit lambda that can be correctly serialized and deseriazed.

This patch modifies code generation so that implicit lambda code
generation is triggered for this case.

Existing testcases must be updated as error message are different when
implicit lambda is detected (although the detected error is the same).

Change-Id: Ic21a268d48630e87e2d766a462c26b889784af99
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/191514
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Jay Arthanareeswaran <jarthana@in.ibm.com>