Bug 479451 - Fix fails in ModuleBuilderTests due to new warning marker
With changes for bug 479451, a new warning marker was introduced for
projects without explicit default encoding.

This marker causes a few fails in JDT ModuleBuilderTests, where tests
expect a specific set of markers and fail due to the new marker. To fix
this, the change ensures test projects that are "imported" from a
template also have an explicit default encoding. To make sure tests will
not interfere with encoding changes reported by CharsetDeltaJob, the
test waits now on this job too.

It also turned out, that DeltaProcessor.updateCurrentDeltaAndIndex() had
an issue reporting "close" deltas on projects with explicit encoding set
- instead, it reported content change. This is also fixed now.

Change-Id: Ia0ec4d19bad2e8495675c4c77d6a7f7e1138e864
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/192239
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>