Fix completions for lambda variables under nested if blocks. Fixes #109
The fix adds a additional guard when trying to build the enclosing
context by collecting if conditions which supposed to have instanceof
previously. But not having the guard condition to only add the if
statement in presence of instanceof or completion node been part of if
statement cause the current issue. The fix add those guard conditions.