Adding Polymorphic method deduction
We are depending on PolymorphicSignature annotation to determine whether
the method is polymorphic or not. But according to
https://docs.oracle.com/javase/specs/jls/se18/html/jls-15.html#jls-15.12.3

A method is signature polymorphic if all of the following are true:

It is declared in the java.lang.invoke.MethodHandle class or the
java.lang.invoke.VarHandle class.
It has a single variable arity parameter (§8.4.1) whose declared type is
Object[].
It is native.
I guess we should deduce PolymorphicSignature annotation based on above
definition

Fixes https://github.com/eclipse-jdt/eclipse.jdt.core/issues/147