Add cmd.-line opt. to take optional warn. prop. file
When the Eclipse Java Language Server analyze a multi-module Maven
project, the lack of a .settings/org.eclipse.jdt.core.prefs file in
some of the modules is silently ignored, as if this file existed but
simply had no contents.

Attempting to recreate this effect when compiling the same project
with the Eclipse Java batch compiler, for instance by using the
org.codehaus.plexus:plexus-compiler-eclipse add-on to the
maven-compiler-plugin, and passing -properties as compilerArgs, is not
possible, because the compiler will emit an error on those projects
which do not have a warnings property file.

This patch introduce an -optprops command-line option, which has the
same semantics as the -properties option, except that missing files
are silently ignored. By specifying this in the root POM, the entire
project may be compiled with the Eclipse Java compiler without having
to commit an empty warnings properties file in each of the
sub-modules.