Move MODULE_TO_LOAD into class final variable / make JRT_URI final
Currently MODULE_TO_LOAD is a static non final variable that is written
on each creation of JrtFileSystem, even though it seems unlikely to
cause issues, it has the feeling of some code-smell.

JRT_URI on the other hand, is non-final but never modified and thus
should be declared final as well.

Relates to https://github.com/eclipse-jdt/eclipse.jdt.core/issues/183