Adding Polymorphic method deduction (#153)
* Adding Polymorphic method deduction

We are depending on PolymorphicSignature annotation to determine whether
the method is polymorphic or not. But according to
https://docs.oracle.com/javase/specs/jls/se11/html/jls-15.html#jls-15.12.3

A method is signature polymorphic if all of the following are true:

It is declared in the java.lang.invoke.MethodHandle class or the
java.lang.invoke.VarHandle class.
It has a single variable arity parameter (§8.4.1) whose declared type is
Object[].
It is native.
I guess we should deduce PolymorphicSignature annotation based on above
definition

Fixes https://github.com/eclipse-jdt/eclipse.jdt.core/issues/147

* Review Discussion
It seems like Java invented an interesting notion of "internal":

The annotation type PolymorphicSignature is package private in MethodHandle, which renders its use outside java.lang.invoke illegal. They could've stopped here.

As the annotation has retention RUNTIME it must be present in MethodHandle.class.

I was surprised to see even RUNTIME as the annotation is intended (as per javadoc) to aid "the Java compiler", but fair enough, likely the VM needs it, too.

But then, in --release mode, MethodHandle.class isn't even used, but some MethodHandle.sig.

Seeing lack of file PolymorphicSignature.sig in ct.sym could be understood as making use of this annotation in user code doubly illegal (why?).

But seeing methods like invokeExact() defined in MethodHandle.sig without any annotation (source code has two), is - weird.

As a result, the tool (java compiler) that is the main client of the annotation cannot see the annotation in --release mode.

If only PolymorphicSignature.sig was missing from ct.sym I would recommend to interpret the name even if it cannot be resolved, but as even the annotation reference in invokeExact() and friends is missing, we don't seem to be allowed to do what the spec requires us to do.

In this light, the find in jls-15.12.3 is precious! :) That's the best possible kludge we can clutch at.

I only see one downside in the deduction strategy: ecj gets more fragile vis-a-vis changes in JLS (that may or may not be announced), whereas reading a manifest annotation would be future proof.

Co-authored-by: Jay Arthanareeswaran <jarthana@in.ibm.com>