Fix type hierarchy for cyclic static imports #269
If two classes import each other with static imports, computing a type
hierarchy can run into problems. In particular:

1. CompilationUnitScope.checkAndSetImports() will try to resolve the
static import in the 1st type.
2. checkAndSetImports() for the 2nd type will try to resolve the static
import.
3. This will result in using the (currently) incomplete scope of the 1st
type.
4. Super types (interfaces in the case of #269) are not resolved
correctly, since the bindings of the 1st type are not yet fully
computed.

This change adjusts CompilationUnitScope.checkAndSetImports() to first
resolve non-static imports and then resolve static imports. This ensures
the statically imported types are resolved as correctly. A test for the
bug is included.

Fixes: #269