Add test for generic method signature of varargs method reference
This test is implemented such that it expects to find no generic
signature for lambdas (like javac) instead of the presently invalid
signature generated by the JDT compiler. The example being tested uses
the java.util.stream.Stream.of method as it is a prominent example of a
varargs method that gets selected in a method reference and thus
produces a synthetic static lambda method with invalid "Signature"
attribute in the generated class file.

Signed-off-by: Christian Femers <christian.femers@t-online.de>