Remove obsoleted MatchLocator.isJarOrClass() code (#422)
The method was added during Java 9 transition as it was not clear how
binary types will be packaged in the JDK.

We don't need to check the path from the class file to determine if it
is a class file from "jar" or "class" or whatever else. If we were able
to compile related code, we should be able to find that binary type - if
not, it is a bug on its own.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.core/issues/422