No match found for method reference in anonymous class #432
The performance improvement done for Eclipse bug 429279 introduces
pruning of method bodies, if a method does not have functional interface
types. Checking for the presence of functional interface types however
does not consider anonymous types defined in the method body.

This change prevents the pruning done for bug 429279, in case a method
defines an anonymous type.

A test is also added for a few example cases.

See:
* https://bugs.eclipse.org/bugs/show_bug.cgi?id=429279
* https://bugs.eclipse.org/bugs/show_bug.cgi?id=566435

Fixes: #432