Refactor look back to be modified only by two methods and add Javadoc
It seemed useful to document what the order of the most recent tokens in
the lookBack array is and what it contains. The change also restricts
what lookBack may contain further by ignoring whitespace and comments.
This seems to be necessary, as many decisions are based on the lookBack
contents and especially comments of the form /*abc*/ in between
'functional' Java language tokens can cause failures (example in
047fa1c03f27c4683fbfea71d4dcd8c732d5eed3)
To the author's best knowledge, no user of lookBack relies on comment or
whitespace tokens being present.