Add a preference to ignore unnamed module for split package
This is to control whether the error 'The package xxx is accessible from
more than one module' will be reported.

In https://openjdk.org/projects/jigsaw/spec/sotms/#the-unnamed-module says

If a package is defined in both a named module and the unnamed module then
the package in the unnamed module is ignored. This preserves reliable configuration
even in the face of the chaos of the class path, ensuring that every module still reads
at most one module defining a given package.

Signed-off-by: Sheng Chen <sheche@microsoft.com>