Improve scanner "when" token disambiguation (#441)
* Add Scanner test that when-token is not detected if used as identifier

With the current implementation, common Java code such as
SomeClass.when(condition).doSomething()
does not compile because the Scanner recognizes it as the "when"
terminal of the JLS 19 patterns-switch-record-patterns "Guard" (see
Section 14.11.1)

* Refactor look back to be modified only by two methods and add Javadoc

It seemed useful to document what the order of the most recent tokens in
the lookBack array is and what it contains. The change also restricts
what lookBack may contain further by ignoring whitespace and comments.
This seems to be necessary, as many decisions are based on the lookBack
contents and especially comments of the form /*abc*/ in between
'functional' Java language tokens can cause failures (example in
047fa1c03f27c4683fbfea71d4dcd8c732d5eed3)
To the author's best knowledge, no user of lookBack relies on comment or
whitespace tokens being present.

* Restrict when token to occur only after identifiers and ')'

This matches the grammar described in
https://docs.oracle.com/javase/specs/jls/se19/preview/specs/patterns-switch-record-patterns-jls.html#jls-14.11.1
and prevents that "SomeClass.when(condition);" is parsed as pattern
switch guard "when" token.

Note that the current rule that TokenNameNotAToken is allowed as well
may still cause problems if single statements get compiled, but this
behavior is required by the current tests.