Java editor cannot resolve import due to classpath filter #485
When using the classpath "including" and "excluding" entry properties,
its possible that a top level package is not listed as a
IPackageFragmentRoot. As a result,
SearchableEnvironment.getModulesDeclaringPackage() fails to recognize
such a package as belonging to the unnamed module. This can result in
Java editor unresolvable type compile errors.

This change adjusts SearchableEnvironment.getModulesDeclaringPackage()
to consider parents of packages in the unnamed module to also belong in
the unnamed module.

See also https://github.com/eclipse-jdt/eclipse.jdt.core/issues/646

Fixes: https://github.com/eclipse-jdt/eclipse.jdt.core/issues/485
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>