Use meta index for reference search
The meta index which was introduced to speedup subtype search could be
used to speed up the reference search as well. When proper qualifier
parameters are given with the pattern, the index selection will use that
information to narrow down the correct indexes to search for.

This change will only add support for using meta index information for
type reference and declaration search for now.