meta-index fix for for move/rename #638
When performing rectorings such as move/rename meta-index doesn't have
correct information, therefore when we do index filtering based on
meta-index results if we find that none of the selected indexes are not
present in meta index, to keep search backward compatible, we return the
original selection to search engine.

On edge cases this could lead to more wider search which should have
stopped by the meta-index, but until we find a proper solution this will
restore the previous behavior.