Enhance the BuildContext with the discovered annotations
Currently one can only get the information that annotations are present
on a file in a BuildContext / CompilationParticipant but sometimes it is
much more interesting to know if a specific annotation is present so one
don't need to scann files unnecsaryly.

This captures the found low-level annotations and adds a new method
based on this BuildContext.hasAnnotations(String) so that
CompilationParticipant can check fi the file is of any interest for
them.

Fix https://github.com/eclipse-jdt/eclipse.jdt.core/issues/674