No call hierarchy for method reference after fixing compile error
SourceIndexer.indexResolvedDocument() doesn't add an index entry for a
method reference with an unresolved method binding. If the expression
becomes resolved later on and so has a valid method binding, e.g. due to
fixing compile errors in another file, SourceIndexer is not triggered
again. So the index for the file becomes incomplete and the method
reference is not found e.g. when computing the call hierarchy of the
referenced method.

This change adds a re-indexing request, if a delta is seen that removes
JDT error markers from a resource. This ensures a source with method
references is re-indexed after compile errors are fixed, so that the
index for the source is in the expected state.

Fixes: #438
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>