Bug 578073 - [18] JSR 269 - Implement Elements.getFileObjectOf(Element) (#668)
The bindings, especially, the binary ones, don't hold the path to the corresponding resource. ClassFileReader, when created, had to be embedded with the path to the same. Once we have that, we can propagate that to the bindings and the same can be used when creating FileObject.

And make sure to tse the FileSystems for reading from Jar files also, which was being used only for JRT. Also note, the test in JavaSearchBugsTests.testBug261722() needed to be adjusted to avoid 
deletion of resources along with the project.