Bad call hierarchy for method in JRE 8 project called by JRE 11 project
During a call hierarchy computation, MatchLocator.initialize() can
create a JavaSearchNameEnvironment with two projects. One is the project
which contains the method or field (the root of the hierarchy), the
other is the project in which MatchLocator looks for references.

The search environment for the two project is created with the combined
classpath entries of both projects. This can result in module conflicts,
e.g. when one project has JRE 8 on the classpath and the other project
has JRE 11 on the classpath. Both JREs contain many types in common, the
JRE 8 will usually be on the classpath and the JRE 11 will usually be on
the module path. Thus types from the JRE 8 will be in the unnamed
module, types from the JRE 11 will be in the respective Java module.

This change adjusts the creation of the JavaSearchNameEnvironment to
take only exported classpath entries and source classpath entries from
the secondary project (the project containing the method or field for
which the call hierarchy is being computed).

Fixes: https://github.com/eclipse-jdt/eclipse.jdt.core/issues/723
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>