Don't use FileSystem just to read zip file entries
Using a FileSystem is not correct way to read zip file entries in
multi-threaded environment, because FileSystem is a global object and
may be already created or closed while we are reading from it, and vice
versa.

Beside this, this change fixes
org.eclipse.jdt.ui.tests.leaks.JavaLeakTest failing because of the
unexpected use of ZipFileSystem on opening JDT editor and compilation
error if compiling JDT with Java 17.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.core/issues/722