Handle JRE container when adding module class path info in Java search
The JRE container on a projects classpath doesn't always have the module
attribute set, despite being on the module path. Due to not considering
this, the fix for #675 introduces a bug - types in the JRE modules are
not resolved correctly.

The fix for #740 disables the fix for #675 for classpath containers to
prevent this bug from occurring.

This change ensures classpath containers with system modules (such as
the JRE container) have their classpath module info attached, while the
module info is not attached for other containers without the module
classpath attribute. This ensures the bug in #675 is fixed also for
containers on the classpath.

Fixes: https://github.com/eclipse-jdt/eclipse.jdt.core/issues/748

Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>