Call hierarchy incomplete for method of private nested type
An incomplete call hierarchy is computed given the following conditions:

1. A private nested class implements an interface.
2. The interface has a method with a nested type.
3. The call hierarchy starts at the private nested class level.

This is a regression from the fix for:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=431357

In particular the fix uses MethodBinding.parameters, which is empty for
the callee method in the described scenario.

A MethodBinding with empty parameters is computed due to the following
sequence:

1. The method caller is not able to see the private private nested class
with the callee method, resulting in creating a ProblemReferenceBinding
for the nested private class.
2. MatchLocator.getMethodBinding(MethodPattern) fails to retrieve a
MethodBinding via MatchLocator.getMethodBinding0(MethodPattern), due to
the ProblemReferenceBinding.
3. getMethodBinding() resorts to parsing source code, but retrieves a
MethodBinding with unresolved parameters.

This change adjusts the part of getMethodBinding() which resorts to
reading a source file. The logic for binary methods is re-used also for
source methods: if a ProblemReferenceBinding offers a closest match, we
use it to retrieve the MethodBinding.

Fixes: #821
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>