Field assignment to anonymous type breaks selection
Assigning an anonymous inner class to a class field results in a
dysfunctional selection (e.g. open declaration doesn't work). Normally,
AssistParser.buildInitialRecoveryState() will ensure completeness of the
code block in which the selection is evaluated.

An assignment expression is not handled in buildInitialRecoveryState(),
and so the context in which the selection is evaluated is incomplete.
This can lead to the dysfunctional selection.

This change adjusts buildInitialRecoveryState() to handle assignment
expressions, so that selections work as expected.

Fixes: #860
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>